Synthesizer 3D phase effects implemented
So far the stereo information was only implemented as a different
amplitude of the left and the right channel. However, what is even more
important for stereo impression is the phase shift resulting from the
different times at which the sound waves arrive at our ears. Now the
synthesizer takes this effect into account by phase-shifting one of the
two sine waves. The result is a much more 3D-like impression when
playing.