Corrected tick marks in the ZoomedSpectrum
In the zoomed spectrum where the rectangle is moving we have a couple of
vertical gray lines in the background which should display distances of
10 cents each. This was not in coincidence with the measured value. The
markers were set in a hard way, however, they should be set according to
the width of the array that is displayed. I modified the grid lines and
inserted also small tick marks for single cents.