Possible critical bug resolved
We had a number of crashes on the iPad and one crash on Android. Today I
observed a crash unter Linux in the debug mode. It occured in
KeyRecognized in the SignalAnalyzer. As a possible origin it could be
that another thread cancels the key count statistics just in the moment
when the loop is executed to find the maximum. I found that this could
happen when the KeyRecognizer calls the SignalAnalyzer via callback
function. To prevent the crashes I added a mutex.