Simple thread handler: Assigning thread names in Linux debug mode
The SimpleThreadHandler is modified in such a way that it can handle
thread names. This simplifies mutex debugging in the Qt-creator as the
debugger shows the name of the thread (e.g. "Synthesizer"). However, Qt
does not offer a platform-independent way to do so, so the function
setThreadName is implemented with macros in such a way that it becomes
only active under Linux (not Android) in the Qt-Debug mode. In this case
it calls the function of the underlying pthread-implementation.