Bug in the audorecorder fixed
If the input level was overdriven the audiorecorder did not down-
regulate the input level sufficiently. Instead a residual clipping
remained. The mistake was that we downregulated only when the intensity
was above 0.9. The intensity has the maximum 1 for a perfect rectangular
wave. But it has the intensity 1/2 for a sine wave. This means that we
have to work with a threshold below 1/2. I chose 0.45. Now the down-
regulation works properly.