Bug in the synthesizer removed
In the synthesizer there was a bug caused by the phase shift between the
two stereo channels. It became effective only in the lower half of the
keyboard and caused a disturbing sound at the onset depending on the
memory content. Therefore, it was not audible on all devices. An
additional condition checking the array bounds corrects the bug.
Moreover, we removed a MIDI pedal switching option that we used once for
testing.