Bug in the stroboscopic display resolved
The strobe showed wrong values in the bass. This was caused by a
discrepancy between the computed fundamental frequency on the basis of
the partials and the actual fundamental frequency of the lowest peak.
This discrepancy occurs only in the case of copper-wounded strings and
is not yet understood. However, taking the correct fundamental frequency
the bug in the strobe can be removed. Yet we have to understand this
phenomenon more carefully in the future.