Added support for Windows dll build
This is required if the core is built dynamically. The tp3log
symbols won't be exported and are missing for all other parts.
The tp3log.dll is a new necessary dependency on windows desktop
builds. WinRT uses a static core and can thus use the static
tp3log.
The configuration is done in .qmake.conf.