Fixes #50: Added static config and tp3LogDLL Flag
Added a static configuration pri that does not depend on running tests.
This is required for adding a global config also to subprojects/modules.
The tests (e. g. for searching qwt) are performed in a dynamic pri
that is only included in a .pro file.
The complete change is a workaround for Qt creating the default log
file in / when no root dir is set.

Fixed the tp3LogDLL flag in the core dependencies in a shared core but
static tp3Log build.